requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.0
billiard==4.2.0
blinker==1.4
#Brlapi
celery==5.3.5
certifi==2022.6.15
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.5
#command-not-found
cryptography==3.4.8
#cupshelpers
dbus-python==1.2.18
decorator==5.1.1
defer
distro==1.7.0
distro-info==1.1+ubuntu1
Django==4.2
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.22
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.17.3
future==0.18.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.6.4
ipython==8.16.1
jedi==0.19.1
jeepney==0.8.0
jsonschema==4.6.0
keyring==23.8.2
kombu==5.3.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.22.0
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.10.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.2
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.46
python-dotenv==0.20.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.1
reportlab==3.6.11
requests==2.27.1
SecretStorage==3.3.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
systemd-python==235
texttable==1.6.4
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.8
websocket-client==1.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
